consumer/requirements.txt
pandas
pyarrow
kafka-python
fastavro
python-dotenv